+++ frontend/src/components/Header.js
@@ -2,8 +2,9 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { Tooltip, IconButton, Button, createTheme } from '@mui/material';
-import RestartAlt from '@mui/icons-material/RestartAlt';
-
+import DriveFolderUploadIcon from '@mui/icons-material/DriveFolderUpload';
+import SystemUpdateAltIcon from '@mui/icons-material/SystemUpdateAlt';
+import RestartAltIcon from '@mui/icons-material/RestartAlt';
 /*
     <div className="header" style={{ display: 'flex', alignItems: 'center', justifyContent: 'flex-start', gap: '20px', margin: '0', padding: '20px', height: '40px' }}>
         <Link to="/test" className="nav-link">Test</Link>
@@ -18,7 +19,43 @@ const theme = createTheme({
   },
 });
 
+export function sendCommandToMatterbridge(command, param) {
+  console.log('sendCommandToMatterbridge:', command, param);
+  // Send a POST request to the Matterbridge API
+  fetch(`/api/command/${command}/${param}`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  })
+  .then(response => {
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+    return response.json();
+  })
+  .then(json => {
+    console.log('Command sent successfully:', json);
+  })
+  .catch(error => {
+    console.error('Error sending command:', error);
+  });
+}
+
 function Header() {
+  // Define the function that sends the "restart" command
+  const handleAddPluginClick = () => {
+    sendCommandToMatterbridge('addplugin','xxxx');
+  };
+
+  const handleUpdateClick = () => {
+    sendCommandToMatterbridge('update','now');
+  };
+
+  const handleRestartClick = () => {
+    sendCommandToMatterbridge('restart','now');
+  };
+
   return (
     <div className="header">
       <img src="matterbridge 64x64.png" alt="Matterbridge Logo" style={{ height: '30px' }} />
@@ -29,7 +66,9 @@ function Header() {
         <Link to="/settings" className="nav-link">Settings</Link>
       </nav>
       <div className="header" style={{ flex: 1, display: 'flex', flexDirection: 'row', justifyContent: 'flex-end' }}>
-        <Tooltip title="Restart matterbridge"><Button theme={theme} color="primary" variant="contained" size="small" endIcon={<RestartAlt />} style={{ color: '#ffffff' }}>Restart</Button></Tooltip>        
+        <Tooltip title="Add plugin"><Button theme={theme} color="primary" variant="contained" size="small" endIcon={<DriveFolderUploadIcon />} style={{ color: '#ffffff' }} onClick={handleAddPluginClick}>Add plugin</Button></Tooltip>        
+        <Tooltip title="Update matterbridge"><Button theme={theme} color="primary" variant="contained" size="small" endIcon={<SystemUpdateAltIcon />} style={{ color: '#ffffff' }} onClick={handleUpdateClick}>Update</Button></Tooltip>        
+        <Tooltip title="Restart matterbridge"><Button theme={theme} color="primary" variant="contained" size="small" endIcon={<RestartAltIcon />} style={{ color: '#ffffff' }} onClick={handleRestartClick}>Restart</Button></Tooltip>        
       </div>
     </div>
   );